Fixes issue when BentoBox blocks other worlds
The previous implementation was blocking shops in worlds, where BentoBox does not operate.

This implementation will allow only island members to create the shops. 
BentoBox offers to create a custom "Flag" for an island, that could allow the owner to modify access to creation and using these shops. However, I do not know if it is necessary, and it will overcomplicate things :)